implement efficient modified file transfer (rsync algorithm) 🎉
Previously, we only implemented the rsync *protocol*, but always transferred all
changed files in full.

Now, we also implement the rsync *algorithm*, meaning we’ll match existing
chunks of data (on the receiver side) and only send what needs to be sent.

This implementation is inefficient, but apparently correct.
In a next step, we can make it more efficient by using fewer Read() calls,
and updating the checksum instead of re-calculating it.