receiver: write files to disk 🎉
File contents, mtime, uid, and gid will be preserved.
Symbolic links and special device nodes are supported.

On Linux and mac, files are written atomically using
https://pkg.go.dev/github.com/google/renameio/v2.
On Windows, atomic replacement is not possible,
but the code still writes to a temporary file and moves it.

Next up: generating checksums to transfer deltas only (rsync algorithm).