Fix symlink race condition (TOCTOU): use O_NOFOLLOW to open files
The rsync vulnerability CVE-2024-12747 also affected gokrazy/rsync
(before this commit): An attacker could replace a regular file
(= contents will be transferred) with a symlink pointing outside
of the specified directory.

To reproduce the issue, run TestReceiverSymlinkTraversal with the
following change to rsyncd/rsyncd.go applied to simulate an attack:

    --- i/rsyncd/rsyncd.go
    +++ w/rsyncd/rsyncd.go
    @@ -373,6 +375,16 @@ func (s *Server) HandleConn(module Module, rd io.Reader, crd *countingReader, cw

     	s.logger.Printf("file list sent")

    +	// HACK: swap out the passwd file with a symlink to /etc/passwd
    +	if err := os.Remove(filepath.Join(module.Path, "passwd")); err != nil {
    +		return err
    +	}
    +	if err := os.Symlink("../passwd", filepath.Join(module.Path, "passwd")); err != nil {
    +		return err
    +	}
    +	s.logger.Printf("HACK: swapped passwd file for symlink")
    +
     	// Sort the file list. The client sorts, so we need to sort, too (in the
     	// same way!), otherwise our indices do not match what the client will
     	// request.

Before this commit, the test would transfer the "secret" contents;
after this commit, the test fails with "no such file or directory".