fix transferring individual files after os.OpenRoot+landlock
Both defense mechanisms were getting in the way:

Whereas filepath.WalkDir() transparently handles calling
WalkDir on a file, the new approach of os.OpenRoot+WalkDir
requires that the argument to OpenRoot is a directory.

Similarly, we passed our sources directly to landlock.RODirs(),
but we should tease the names apart into files and directories,
so that we can call landlock.RODirs() and landlock.ROFiles().